encapsulated MultiLevelPtr.cs
