+++ SimpleMem/MultiLevelPtr.cs
@@ -25,6 +25,16 @@ namespace SimpleMem;
 public class MultiLevelPtr
 {
     /// <summary>
+    ///  Base address of the pointer chain
+    /// </summary>
+    public IntPtr Base { get; set; }
+    /// <summary>
+    ///  Optional list of offsets containing pointer offsets (from the provided base).
+    /// </summary>
+    public List<IntPtr> Offsets { get; set; } = [];
+
+    #region Constructor
+    /// <summary>
     ///  <inheritdoc cref="MultiLevelPtr" />
     /// </summary>
     /// <param name="lpBaseAddress">The address the pointer starts from. This is almost always the ModuleBaseAddress.</param>
@@ -70,7 +80,7 @@ public class MultiLevelPtr
     {
         this.Base = baseMlPtr.Base;
 
-        if (baseMlPtr.Offsets.Any())
+        if (baseMlPtr.Offsets.Count != 0)
         {
             foreach (var offset in baseMlPtr.Offsets)
             {
@@ -100,7 +110,7 @@ public class MultiLevelPtr
     {
         this.Base = baseMlPtr.Base;
 
-        if (baseMlPtr.Offsets.Any())
+        if (baseMlPtr.Offsets.Count != 0)
         {
             foreach (var offset in baseMlPtr.Offsets)
             {
@@ -143,7 +153,7 @@ public class MultiLevelPtr
     public MultiLevelPtr(IntPtr[] pointers)
     {
         this.Base = pointers[0];
-        this.Offsets = pointers[1..];
+        this.Offsets = new(pointers[1..]);
     }
 
     /// <summary>
@@ -165,33 +175,11 @@ public class MultiLevelPtr
         this.Base = new IntPtr(pointers[0]);
         this.Offsets = ConvertLongs(pointers[1..]);
     }
-
-    /// <summary>
-    ///  Base address of the pointer chain
-    /// </summary>
-    public IntPtr Base { get; set; }
-    /// <summary>
-    ///  Optional list of offsets containing pointer offsets (from the provided base).
-    /// </summary>
-    public IList<IntPtr> Offsets { get; set; } = new List<IntPtr>();
-
-    /// <inheritdoc />
-    public override string ToString()
-    {
-        var sb = new StringBuilder($"MultiLevelPtr(Base={this.Base:X}, Offsets=[");
-        foreach (var offset in this.Offsets)
-        {
-            sb.Append($"{offset:X}, ");
-        }
-
-        sb.Remove(sb.Length - 2, 2);
-        sb.Append("])");
-        return sb.ToString();
-    }
+    #endregion
 
     private static List<IntPtr> ConvertInts(int[] ints)
     {
-        var n = new List<IntPtr>(ints.Length);
+        List<IntPtr> n = new(ints.Length);
         foreach (int i in ints)
         {
             n.Add(new IntPtr(i));
@@ -202,7 +190,7 @@ public class MultiLevelPtr
 
     private static List<IntPtr> ConvertLongs(long[] ints)
     {
-        var n = new List<IntPtr>(ints.Length);
+        List<IntPtr> n = new(ints.Length);
         foreach (long i in ints)
         {
             n.Add(new IntPtr(i));
@@ -210,103 +198,16 @@ public class MultiLevelPtr
 
         return n;
     }
-}
 
-/// <summary>
-///  Class for conveniently definining multi-level pointers.
-///  Multi-level pointers are used for obtaining values even
-///  after restarts of programs (unlike a single,
-///  non-static memory address).
-///  <example>
-///   Consider a game called "MyGame" with a module of "MyGame.exe".
-///   "MyGame.exe" is given a memory address that changes with each
-///   launch of the program, but all of the values desired (such as
-///   gold, points, exp, etc.) lay the same distance away in memory
-///   from the module. Say the offset for gold is 0xC and exp is 0xD.
-///   A MultiLevelPtr can be created from the "base address" of the module
-///   with the offsets being 0xC for gold and 0xD for exp. Assuming the
-///   base address and offsets are correct, the desired values will
-///   always be returned.
-///  </example>
-/// </summary>
-/// <typeparam name="T">The expected type resolved from the MultiLevelPtr</typeparam>
-public class MultiLevelPtr<T> : MultiLevelPtr where T : struct
-{
-    /// <inheritdoc />
-    public MultiLevelPtr(IntPtr lpBaseAddress, params IntPtr[] offsets) : base(lpBaseAddress, offsets) { }
-
-    /// <inheritdoc />
-    public MultiLevelPtr(IntPtr lpBaseAddress, params int[] offsets) : base(lpBaseAddress, offsets) { }
-
-    /// <inheritdoc />
-    public MultiLevelPtr(MultiLevelPtr baseMlPtr, params int[] offsets) : base(baseMlPtr, offsets) { }
-
-    /// <inheritdoc />
-    public MultiLevelPtr(MultiLevelPtr baseMlPtr, params long[] offsets) : base(baseMlPtr, offsets) { }
-
-    /// <inheritdoc />
-    public MultiLevelPtr(long lpBaseAddress, params int[] offsets) : base(lpBaseAddress, offsets) { }
-
-    /// <inheritdoc />
-    public MultiLevelPtr(IntPtr[] pointers) : base(pointers) { }
-
-    /// <inheritdoc />
-    public MultiLevelPtr(int[] pointers) : base(pointers) { }
-
-    /// <inheritdoc />
-    public MultiLevelPtr(long[] pointers) : base(pointers) { }
-}
-
-/// <summary>
-/// Extensions for Memory
-/// </summary>
-public static class MemoryExtensions
-{
-    /// <summary>
-    /// Extension that serves as a wrapper for mem.ModuleBaseAddress + offset.
-    /// Useful for storing lots of static offsets that derive directly from
-    /// the ModuleBaseAddress (these are not the same as pointers, although they are similar).
-    /// </summary>
-    /// <param name="mem">The memory instance in which the base address exists</param>
-    /// <param name="offset">Any address</param>
-    /// <returns></returns>
-    public static IntPtr StaticOffset(this Memory mem, int offset) => mem.ModuleBaseAddress + offset;
-    /// <inheritdoc cref="StaticOffset(SimpleMem.Memory,int)"/> 
-    public static IntPtr StaticOffset(this Memory mem, long offset) => new((long)mem.ModuleBaseAddress + offset);
-}
-
-/// <summary>
-/// Extensions for MultiLevelPtr
-/// </summary>
-public static class PointerExtensions
-{
-    /// <summary>
-    ///  Reads the value, in memory, of this <see cref="MultiLevelPtr{T}" />. This extension saves
-    ///  a separate call to <see cref="Memory.ReadAddressFromMlPtr" />.
-    /// </summary>
-    /// <param name="mlPtr">The <see cref="MultiLevelPtr{T}" /> to read the value from</param>
-    /// <param name="mem">The <see cref="Memory" /> instance in which this value lays</param>
-    /// <returns></returns>
-    public static T ReadValue<T>(this MultiLevelPtr<T> mlPtr, Memory mem) where T : struct => mem.ReadValueFromMlPtr(mlPtr);
-
-    /// <summary>
-    /// Writes the value in the same fashion as <see cref="Memory.WriteMemory{T}"/>
-    /// </summary>
-    public static int WriteValue<T>(this MultiLevelPtr<T> mlPtr, Memory mem, T val) where T : struct
-        => mem.WriteMemory(mlPtr.GetAddress(mem), val);
-
-    /// <summary>
-    /// Writes the bytes to memory at the address resolved from the MultiLevelPtr.
-    /// </summary>
-    /// <returns>The number of bytes written</returns>
-    public static int WriteBytes(this MultiLevelPtr mlPtr, Memory mem, params byte[] bytes) =>
-        mem.WriteMemory(mlPtr.GetAddress(mem), bytes);
-
-    /// <summary>
-    ///  Gets the address resolved from the given mlPtr.
-    /// </summary>
-    /// <param name="mlPtr">The <see cref="MultiLevelPtr" /> to read the address from</param>
-    /// <param name="mem">The <see cref="Memory" /> instance in which this address lays</param>
-    /// <returns>A IntPtr containing the address, if found.</returns>
-    public static IntPtr GetAddress(this MultiLevelPtr mlPtr, Memory mem) => mem.ReadAddressFromMlPtr(mlPtr);
+    public override string ToString()
+    {
+        StringBuilder sb = new($"MultiLevelPtr(Base={this.Base:X}, Offsets=[");
+        foreach (nint offset in this.Offsets)
+        {
+            sb.Append($"{offset:X}, ");
+        }
+        sb.Remove(sb.Length - 2, 2);
+        sb.Append("])");
+        return sb.ToString();
+    }
 }
